requirements.txt
# Core
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Image processing
opencv-python==4.8.0.74
Pillow==9.5.0

# Machine Learning / Deep Learning
scikit-learn==1.3.0

# Visualization  
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Web framework
streamlit==1.25.0

# Configuration
pyyaml==6.0.1

# Utilities
jupyter==1.0.0
